Make use of enummapset
The collection wrappers used internally are more limited compared to
what is available in IntMap/IntSet. What the try to do is add
different types on top of IntMap and IntSet. I have yet to figure out
how I can make collection functions such as restrictKeys that need
both types, to be accepted by GHC.

Instead, I think it is easier to just use enummapset which provides a
more complete set of functions and we can avoid trying to keep this
updated.

A module Compiler.Hoopl.Compat is added that makes it easier to use
this with existing code without having to make changes except
importing that extra module.

I have made some basic profiling and I could not find any noticeable
difference, it mostly disappeared in the noise. If there is any
indication, it seemed that the new way may be very slightly faster,
but it is as mentioned hard to tell.

I also disabled Safe language for now as enummapset is not Safe so it
will not build with that.